add T09

+++ T09/src/main/java/com/github/vitorialuz229/poo/t09/circulo.java
@@ -1,7 +1,9 @@
 package com.github.vitorialuz229.poo.t09;
 
+import java.util.ArrayList;
+
 public class Circulo {
-    private short raio;
+    private int raio;
 
     private Ponto centro = new Ponto();
     
@@ -12,7 +14,7 @@ public class Circulo {
     
 public void transladar(int dx, int dy) { 
 
-    
+
 }
